+++ src/Vectron.Extensions.Logging.TextBlock/TextBlockLoggerFormatterOptionsChangeTokenSource.cs
@@ -1,5 +1,6 @@
 using Microsoft.Extensions.Logging.Configuration;
 using Microsoft.Extensions.Options;
+using Vectron.Extensions.Logging.TextBlock.Formatters;
 using Vectron.Extensions.Logging.TextBlock.Internal;
 
 namespace Vectron.Extensions.Logging.TextBlock;
